Move parametric scenario def in it's own file
