+++ utils/scripts/compute_impacted_scenario.py
@@ -118,6 +118,8 @@ def main():
             elif file.startswith("utils/"):
                 if file.startswith("utils/interfaces/schemas"):
                     scenarios_groups.add(ScenarioGroup.END_TO_END.value)
+                elif file == "utils/_context/_scenarios/parametric.py":
+                    scenarios_groups.add(ScenarioGroup.PARAMETRIC.value)
                 else:
                     scenarios_groups.add(ScenarioGroup.ALL.value)
